Remove `DecorationManagerBuilder`, use `DecorationManagerAdapter` in `DecorationBasicManager` instead.
