+++ src/miral/CMakeLists.txt
@@ -33,6 +33,7 @@ add_library(miral-internal STATIC
                                          window_info_defaults.h
     window_specification_internal.cpp    window_specification_internal.h
     decoration_window_state.cpp
+    decoration_manager_adapter.cpp
 )
 
 # Already implied by the linker's symbol version script, but can avoid accidents
@@ -90,7 +91,6 @@ add_library(miral-external OBJECT
     zone.cpp                            ${miral_include}/miral/zone.h
     decorations.cpp                     ${miral_include}/miral/decorations.h
     custom_decorations.cpp		${miral_include}/miral/custom_decorations.h
-    decoration_manager_builder.cpp	${miral_include}/miral/decoration_manager_builder.h
     decoration_basic_manager.cpp	${miral_include}/miral/decoration_basic_manager.h
     decoration_adapter.cpp		${miral_include}/miral/decoration_adapter.h
     decoration.cpp
