+++ src/miral/decoration_manager_adapter.cpp
@@ -1,0 +1,43 @@
+/*
+ * Copyright © Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 or 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "decoration_manager_adapter.h"
+#include "mir/shell/decoration/manager.h"
+
+#include <memory>
+
+miral::DecorationManagerAdapter::DecorationManagerAdapter() = default;
+miral::DecorationManagerAdapter::~DecorationManagerAdapter() = default;
+
+void miral::DecorationManagerAdapter::init(std::weak_ptr<mir::shell::Shell> const& shell)
+{
+    on_init(shell);
+}
+
+void miral::DecorationManagerAdapter::decorate(std::shared_ptr<mir::scene::Surface> const& surface)
+{
+    on_decorate(surface);
+}
+
+void miral::DecorationManagerAdapter::undecorate(std::shared_ptr<mir::scene::Surface> const& surface)
+{
+    on_undecorate(surface);
+}
+
+void miral::DecorationManagerAdapter::undecorate_all()
+{
+    on_undecorate_all();
+}
